+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Initial boilerplate.
 - Add readme
 - Add tests
+- Add --force parameter to add